+++ src/Symfony/Bundle/TwigBundle/DependencyInjection/Compiler/TwigLoaderPass.php
@@ -40,9 +40,24 @@ class TwigLoaderPass implements CompilerPassInterface
             $container->setAlias('twig.loader', key($loaderIds));
         } else {
             $chainLoader = $container->getDefinition('twig.loader.chain');
-            foreach (array_keys($loaderIds) as $id) {
-                $chainLoader->addMethodCall('addLoader', array(new Reference($id)));
+
+            $prioritizedLoaders = array();
+
+            foreach ($loaderIds as $id => $tags) {
+                foreach ($tags as $tag) {
+                    $priority = isset($tag['priority']) ? $tag['priority'] : 0;
+                    $prioritizedLoaders[$priority][] = $id;
+                }
             }
+
+            krsort($prioritizedLoaders);
+
+            foreach ($prioritizedLoaders as $loaders) {
+                foreach ($loaders as $loader) {
+                    $chainLoader->addMethodCall('addLoader', array(new Reference($loader)));
+                }
+            }
+
             $container->setAlias('twig.loader', 'twig.loader.chain');
         }
     }