+++ src/Symfony/Bundle/TwigBundle/Tests/Extension/LegacyAssetsExtensionTest.php
@@ -15,7 +15,7 @@ use Symfony\Bundle\TwigBundle\Extension\AssetsExtension;
 use Symfony\Bundle\TwigBundle\Tests\TestCase;
 use Symfony\Component\Routing\RequestContext;
 
-class AssetsExtensionTest extends TestCase
+class LegacyAssetsExtensionTest extends TestCase
 {
     /**
      * @dataProvider provideGetGetAssetUrlArguments