+++ src/Symfony/Component/Templating/Tests/Helper/LegacyCoreAssetsHelperTest.php
@@ -13,12 +13,14 @@ namespace Symfony\Component\Templating\Tests\Helper;
 
 use Symfony\Component\Templating\Helper\CoreAssetsHelper;
 
-class CoreAssetsHelperTest extends \PHPUnit_Framework_TestCase
+class LegacyCoreAssetsHelperTest extends \PHPUnit_Framework_TestCase
 {
     protected $package;
 
     protected function setUp()
     {
+        $this->iniSet('error_reporting', -1 & ~E_USER_DEPRECATED);
+
         $this->package = $this->getMock('Symfony\Component\Templating\Asset\PackageInterface');
     }